feat(SplineWidget): Improve the widget behavior
In behavior.js:
- Make private methods actually private (updateHandlesSize, addPoint,
getHoveredHandle)
- Use `macro.setGet` to add public APIs around the model properties, and
remove manual setters (sethandleSizeInPixels and setFreehandMinDistance)
- Override setHandleSizeInPixels and setResolution to update the model
representations
- Lose the focus when ending the interaction, to avoid still placing
points, for example when pressing Enter.
- Remove updateResolution which hardcoded a non-standard behavior when
pressing "Shift" using the `renderPoly` model property. Instead, the
setResolution API can be used by the application with a custom keyDown
listener.
- Add missing event callbacks in grab/loseFocus

In the example:
- Improve the control Panel to interact with most spline parameters

Notes:
- color, scale and visibility of the handles should probably be a property
of the viewWidgets and not of the state.
- there are confusions between the model properties on the widget
(widget factory) and the behavior (view widget). The former can be set to
define the default values when instantiating a view widget, but the later
should be used to edit the properties of a view widget instance.